+++ src/components/combos/ComboEdit.jsx
@@ -1,16 +1,14 @@
 import React, { useContext, useEffect } from "react";
 import ComboContext from "../../Context/ComboContext";
 import OutlinedInput from '@mui/material/OutlinedInput';
-import InputLabel from '@mui/material/InputLabel';
 import MenuItem from '@mui/material/MenuItem';
-import FormControl from '@mui/material/FormControl';
 import ListItemText from '@mui/material/ListItemText';
 import Select from '@mui/material/Select';
 import Checkbox from '@mui/material/Checkbox';
 import { useParams } from "react-router-dom";
 
 export const ComboEdit = () => {
-const { formValues, onChange, errors, setErrors, combo, getCombo, updateCombo, services } = useContext(ComboContext);
+const { formValues, onChange, errors, setErrors, getCombo, updateCombo, services, MenuProps } = useContext(ComboContext);
 const selectedServiceIds = formValues.service_id || []; // Inicializar como un arreglo vacío si es undefined
 
 let { id } = useParams();
@@ -23,16 +21,6 @@ useEffect(() => {
   fetchCombo();
 }, [id]); 
 
-const ITEM_HEIGHT = 48;
-const ITEM_PADDING_TOP = 8;
-const MenuProps = {
-PaperProps: {
-style: {
-maxHeight: ITEM_HEIGHT * 4.5 + ITEM_PADDING_TOP,
-width: 250,
-},
-},
-};
 
 return (
 <div className="mt-12">
